MOSC-2716 / Implementar processo de LOGs para operações críticas.

+++ app/Http/Controllers/GovernancaController.php
@@ -2,13 +2,17 @@
 
 namespace App\Http\Controllers;
 
+use App\Services\AuditService;
 use App\Models\Osc\Governanca;
 use App\Services\Osc\GovernancaService;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Symfony\Component\HttpFoundation\Response;
 
 class GovernancaController extends Controller
 {
+    private $auditService;
+
     private $service;
 
     /**
@@ -19,6 +23,7 @@ class GovernancaController extends Controller
     public function __construct(GovernancaService $_service)
     {
         $this->service = $_service;
+        $this->auditService = new AuditService();
     }
 
     public function get($id)
@@ -58,8 +63,18 @@ class GovernancaController extends Controller
         try {
             $dados = $request->all();
 
+            $entidade = $this->service->store($dados);
+
+            if (!$entidade)
+            {
+                return response()->json(['Resposta' => 'Objeto não encontrado!'], Response::HTTP_OK);
+            }
+
+            $usuario = Auth::user();
+            $this->auditService->auditar('novoGovernanca', 'Governanca', $entidade->id_dirigente, $usuario->id_usuario, 'criado', $entidade, $request->ip(), $entidade->id_osc);
+
             //Retorna novo registro
-            return response()->json($this->service->store($dados), Response::HTTP_OK);
+            return response()->json($entidade, Response::HTTP_OK);
         }
         catch (\Exception $e) {
             return $e->getMessage();
@@ -70,25 +85,40 @@ class GovernancaController extends Controller
         try {
             $dados = $request->all();
 
-            $governanca = $this->service->update($id, $dados);
+            $dados_old = $this->service->get($id);
 
-            if ($governanca)
+            $entidade = $this->service->update($id, $dados);
+
+            if (!$entidade)
             {
-                return response()->json(['Resposta' => 'Governança atualizada com sucesso!'], Response::HTTP_OK);
+                return response()->json(['Resposta' => 'Objeto não encontrado!'], Response::HTTP_OK);
             }
 
-            return $governanca;
+            $usuario = Auth::user();
+            $this->auditService->auditar('updateGovernanca', 'Governanca', $id, $usuario->id_usuario, $dados_old, $dados, $request->ip(), $dados_old->id_osc);
+
+            return response()->json($dados, Response::HTTP_OK);
         }
         catch (\Exception $e) {
             return $e->getMessage();
         }
     }
 
-    public function delete($id_governanca) {
+    public function delete($id_governanca, Request $request) {
         try {
+            $dados_old = $this->service->get($id_governanca);
+
+            if (!$dados_old)
+            {
+                return response()->json(['Resposta' => 'Objeto não encontrado!'], Response::HTTP_OK);
+            }
+
             if ($this->service->delete($id_governanca))
             {
-                return response()->json(['Resposta' => 'Governança deletada com sucesso!'], Response::HTTP_OK);
+                $usuario = Auth::user();
+                $this->auditService->auditar('deleteGovernanca', 'Governanca', $id_governanca, $usuario->id_usuario, $dados_old, 'deletado', $request->ip(), $dados_old->id_osc);
+
+                return response()->json(['Resposta' => 'Governanca deletado com sucesso!'], Response::HTTP_OK);
             }
         }
         catch (\Exception $e) {
